Address code review issues #6
