+++ src/modules/complianceengine/src/lib/payload.schema.json
@@ -207,6 +207,9 @@
           "$ref": "procedures/NetworkInterface.schema.json#/definitions/audit"
         },
         {
+          "$ref": "procedures/NetworkInterfaceFlag.schema.json#/definitions/audit"
+        },
+        {
           "$ref": "procedures/NoDuplicateEntries.schema.json#/definitions/audit"
         },
         {
@@ -341,6 +344,9 @@
         },
         {
           "$ref": "procedures/NetworkInterface.schema.json#/definitions/remediation"
+        },
+        {
+          "$ref": "procedures/NetworkInterfaceFlag.schema.json#/definitions/remediation"
         },
         {
           "$ref": "procedures/NoDuplicateEntries.schema.json#/definitions/remediation"
