+++ src/modules/complianceengine/tests/CMakeLists.txt
@@ -26,7 +26,7 @@ set(PROCEDURES
     procedures/UniqueGroupIdTest.cpp
     procedures/GsettingsValueTest.cpp
     procedures/MountPointExistsTest.cpp
-    procedures/NetworkInterfaceTest.cpp
+    procedures/NetworkInterfaceFlagTest.cpp
     procedures/MtaLocalOnlyTest.cpp
     procedures/NoDuplicateEntriesTest.cpp
     procedures/NoWorldWritableFilesTest.cpp
