+++ src/modules/complianceengine/tests/MockContext.h
@@ -19,6 +19,7 @@ struct MockContext : public ComplianceEngine::ContextInterface
 {
     MOCK_METHOD(ComplianceEngine::Result<std::string>, ExecuteCommand, (const std::string& cmd), (const, override));
     MOCK_METHOD(ComplianceEngine::Result<std::string>, GetFileContents, (const std::string& filePath), (const, override));
+    MOCK_METHOD(ComplianceEngine::Result<std::vector<ComplianceEngine::InterfaceInfo>>, GetNetworkInterfaces, (), (const, override));
 
     OsConfigLogHandle GetLogHandle() const override
     {
